requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
tzdata==2022.6
wincertstore==0.2
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
